train_data_generator/transformer/requirements.txt
# Core dependencies
lxml>=4.9.0
Pillow>=10.0.0
pyyaml>=6.0
numpy>=1.24.0

# Image processing
opencv-python>=4.8.0
scipy>=1.11.0